LayoutTests/platform/mac/editing/spelling/autocorrection-blockquote-crash-expected.txt

This test checks that markers are correct when auto correcting in the blockquote. If you type "n" and " ", there should be blue dots under information, but is off by one. 
Note, this test can fail due to user specific spell checking data. If the user has previously dismissed 'notational' as the correct spelling of 'notationl' several times, the spell checker will not provide 'information' as a suggestion anymore. To fix this, remove all files in ~/Library/Spelling.

PASS internals.hasAutocorrectedMarker(document, 0, 1) is true
PASS successfullyParsed is true

TEST COMPLETE

would this 
testinformation 
make a difference?
